Pull changelog generation into dedicated module
